Fix --help
Non-method attributes are shown using dir() when listing available
commands.

	usage: ogcli [-h] [{create,list,restore,rest,send,set,setup}]

'rest' is a class instance of OgRest and must not be
shown when listing available commands.

Method members of OgCLI class are the available commands that ogcli can
execute.
Use inspect module in order to get class members (getmembers) and filter
only those that are a method (ismethod).

	usage: ogcli [-h] [{create,list,restore,send,set,setup}]

Fixes: 0f2d1f1dba4 ("Show all valid commands when running ogcli --help")