create image: fix --repo default ip value
rest.URL is the ip and port of the ogServer like: 192.168.56.10:8888,
specified in ogcli.json.

As long as the host and port is specified in rest.URL, split by the ':'
character and extract the ip.

Fixes OpenGnsys clients failing to execute the command due to invalid
repository IP being fed to the underlying bash scripts.